requirements.txt
Pillow==10.0.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
typing_extensions==4.7.1
wifi-qrcode-generator==1.0.1
